ide: microdrive cycle field is set as uint8_t
Really, it is used as a boolean

+++ hw/ide/microdrive.c
@@ -51,7 +51,7 @@ typedef struct {
 
     uint8_t ctrl;
     uint16_t io;
-    int cycle;
+    uint8_t cycle;
 } MicroDriveState;
 
 /* Register bitfields */
